[문자열 덧셈 계산기] 음수 예외 테스트 및 StringAddCalculator 패키지 레벨로 이동
